+++ framy_generator/lib/builder.dart
@@ -4,6 +4,7 @@ import 'package:framy_generator/json_builder.dart';
 import 'package:framy_generator/json_formatter.dart';
 import 'package:framy_generator/resolvers/model_resolver.dart';
 import 'package:framy_generator/resolvers/preset_resolver.dart';
+import 'package:framy_generator/resolvers/register_riverpod_resolver.dart';
 import 'package:framy_generator/resolvers/theme_resolver.dart';
 import 'package:framy_generator/resolvers/widget_resolver.dart';
 import 'package:source_gen/source_gen.dart';
@@ -40,6 +41,15 @@ Builder presetBuilder(BuilderOptions options) {
     PresetResolver(),
     formatOutput: fixJson,
     generatedExtension: '.preset.framy.json',
+    header: '',
+  );
+}
+
+Builder registerRiverpodBuilder(BuilderOptions options) {
+  return JsonLibraryBuilder(
+    RegisterRiverpodResolver(),
+    formatOutput: fixJson,
+    generatedExtension: '.riverpod.framy.json',
     header: '',
   );
 }