+++ framy_generator/test/resolvers/dummy_classes.dart
@@ -1,5 +1,6 @@
 class Color {
   final int value;
+
   const Color(int value) : value = value & 0xFFFFFFFF;
 }
 
@@ -22,3 +23,17 @@ class Key {}
 class User {
   String firstName;
 }
+
+class Provider<T> {
+  final T Function(dynamic x) create;
+
+  Provider(this.create);
+}
+
+abstract class Built<V extends Built<V, B>, B> {
+  V rebuild(Function(B) updates);
+
+  B toBuilder();
+}
+
+class BuiltUserBuilder {}