+++ test_apps/weight_tracker/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:framy_annotation/framy_annotation.dart';
 import 'package:weight_tracker/app_theme.dart';
 import 'package:weight_tracker/home_page.dart';
@@ -11,10 +12,12 @@ void main() {
 class WeightTrackerApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: AppTheme.themeData,
-      home: HomePage(),
+    return ProviderScope(
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: AppTheme.themeData,
+        home: HomePage(),
+      ),
     );
   }
 }