+++ framy_annotation/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.2]
+
+- Added `FramyRegisterRiverpod` and `FramyUseRiverpod` annotations
+
 ## [0.1.1]
 
 - Publish changes from `0.1.0-dev.2`